backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
python-dotenv==1.0.1
cachetools==5.3.3
pydantic==2.7.4
tenacity==8.3.0
google-generativeai==0.7.2
overpy==0.7
geojson==3.1.0
